requirements.txt
streamlit>=1.40.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.4.0
joblib>=1.3.0
plotly>=5.20.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
pillow
opencv-python
scikit-image
joblib